dev-requirements.txt
appdirs==1.4.3
colorama==0.3.9
coverage==4.4
decorator==4.0.11
Django==1.11.1
djangorestframework==3.6.2
factory-boy==2.8.1
Faker==0.7.11
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
packaging==16.8
pickleshare==0.7.4
prompt-toolkit==1.0.14
py==1.4.33
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.5.0
pytest-django==3.1.2
pytest-ipdb==0.1.dev2
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
win-unicode-console==0.5
